requirements.txt
asgiref==3.7.2
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.3.0
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
idna==3.4
Markdown==3.5
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.7
daphne==4.0.0
mysqlclient~=2.2.0
python-dotenv~=1.0.0
channels-redis~=4.1.0
